+++ src/indicators/maximum.rs
@@ -1,4 +1,3 @@
-use std::f64::INFINITY;
 use std::fmt;
 
 use crate::errors::{Result, TaError};
@@ -42,13 +41,13 @@ impl Maximum {
                 period,
                 max_index: 0,
                 cur_index: 0,
-                deque: vec![-INFINITY; period].into_boxed_slice(),
+                deque: vec![f64::NEG_INFINITY; period].into_boxed_slice(),
             }),
         }
     }
 
     fn find_max_index(&self) -> usize {
-        let mut max = -INFINITY;
+        let mut max = f64::NEG_INFINITY;
         let mut index: usize = 0;
 
         for (i, &val) in self.deque.iter().enumerate() {
@@ -101,7 +100,7 @@ impl<T: High> Next<&T> for Maximum {
 impl Reset for Maximum {
     fn reset(&mut self) {
         for i in 0..self.period {
-            self.deque[i] = -INFINITY;
+            self.deque[i] = f64::NEG_INFINITY;
         }
     }
 }