+++ src/lib.rs
@@ -29,6 +29,7 @@
 //! * Trend
 //!   * [Exponential Moving Average (EMA)](crate::indicators::ExponentialMovingAverage)
 //!   * [Simple Moving Average (SMA)](crate::indicators::SimpleMovingAverage)
+//!   * [Weighted Moving Average (WMA)](crate::indicators::WeightedMovingAverage)
 //! * Oscillators
 //!   * [Relative Strength Index (RSI)](indicators/struct.RelativeStrengthIndex.html)
 //!   * [Fast Stochastic](indicators/struct.FastStochastic.html)