+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+#### Unreleased
+
+* Add Weighted Moving Average (WMA)
+
+
 #### v0.5.0 - 2021-06-27
 
 * [breaking] - get rid of error-chain. ta::Error -> ta::TaError